+++ helpers/parseHashtags.js
@@ -3,7 +3,7 @@ function parseHashtags(text) {
         throw new TypeError('A string was expected');
     }
     
-    var tagRegEx = /[#]+[A-Za-z0-9-_-\u00F0-\u02AF]+/g;
+    var tagRegEx = /[#]+[A-Za-z0-9-_äöüÄÖÜß\u00F0-\u02AF]+/g;
     return text.match(tagRegEx);
  }